backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.2
tzdata==2024.2

